Create the task_file module
Tasks are now stored in a json format, operations perfomed on the tasks
are now done in the task_file module instead of lib.rs as it makes more
sense for them to be methods of the TaskFile struct than free functions.

Rewritten the cli tests to check the output rather than the task file,
as that is already done by the unit tests.