Fix completions not showing up sometimes
The bash script was passing the entire word array as 1 word due to the
double quotes, that's fixed now.

Also removed the check to see if the 4th element of the array is empty
since spaces are not included in the word array, additionally the only
requirement for completing task contents for an edit command is the
position of the cursor, the shell should handle the rest.

And disable completions after `--generate-shell-completions`.